[FIX] web_gantt: group by fields.properties in gantt view
Issue:

Grouping by a field.properties in the gantt view will give a traceback.

Steps to reproduce:

- Go on a project
- Swap to the gantt view
- Create a task in the calendar
- Go back to the list view, click on the task
- Click on the wheel icon and add a property with a value
- Go back to the project and its gantt view
- Group by the property you have added to your task

> traceback

Cause of the Issue:

The fields given in the groupby parameters are added to the
read_specification of the "get_gantt_data" by the "_fetchData" method:
https://github.com/odoo/enterprise/blob/ab29a91a0acda94fbb98a0f84c8685b34a31fbe5/web_gantt/static/src/gantt_model.js#L635-L638
https://github.com/odoo/enterprise/blob/ab29a91a0acda94fbb98a0f84c8685b34a31fbe5/web_gantt/static/src/gantt_model.js#L915-L920
However, as a fields.properties such as task_properties represent a
dictionary of "pseudo field", the "field name" of a given property is of
the form "task_properties.hashKey". The read_specification of the gantt
view will therefore contain a "field_name" of that form. However, the
ORM does not support to read such fields and the code will crash here:
https://github.com/odoo/enterprise/blob/ab29a91a0acda94fbb98a0f84c8685b34a31fbe5/web_gantt/models/models.py#L89
pseudo fields such as "task_properties.12c97bca432np24" are not expected
to be read (this is not supported by the ORM).

Fix:

If we were to format field names in the "_fetchData" method we would be
able to group records by fields.properties in the gantt view. However,
the framework is not ready to plainly support this operation as for
instance other tracebacks will be encountered when you drag and drop
tasks in the gantt view after a group by a given property.
It was therefore decided with the framework team to remove the
fields.properties from the fields appearing in the groupedBy parameters
of the gantt view until these other operations become supported.

opw-3847341

closes odoo/enterprise#61585

X-original-commit: cfb8889062bb415f2baa7ba961c2c40a2dccf6c8
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>