[FIX] sale_subscription: set correct editable attr in list
The "editable" attribute in list views can be set to either "top"
or "bottom". List views are rng validated, and no other value can
be used. However, x2many list inside form views aren't validated,
and a few of them have editable="1". This seems to work, but only
partially: the list is editable indeed, but if the list is multi
page, there'll be issues when adding records (they'll be added in
the last page, not the current one).

This commit thus fixes the few views with incorrect editable
attribute.

opw~3860903

closes odoo/enterprise#61288

Related: odoo/odoo#162937
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>