[FIX] website_delivery_sendcloud: use complete address for pick-up location
Steps to reproduce:
1. Configure Sendcloud shipping with pick-up location
2. Go to website, use the shipping method and select a
pick-up location
3. Go to the backend, confirm the sale order and validate
the delivery
4. Get Error from Sendcloud that the `house_number` is
required

If we check the delivery address, we see that the house
number is not taken into account when creating a new partner,
and the address is incomplete.

This commit adds the house number to the `street` field of the
address to have a complete address.

opw-3864382

closes odoo/enterprise#61396

X-original-commit: c6698d078c06526899ef6cd9e28c283ce6ff0f69
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Ahmad Khanalizadeh (khah) <khah@odoo.com>