[FIX] documents_spreadsheet: display up-to-date spreadsheet thumbnail
steps to reproduce:

- create a new spreadsheet
- change the background color of some cells
  (so that it's well visible in the thumbnail)
- go back to Documents with the breadcrumb
- open the spreadsheet again
- change the background color
- go back again to Documents with the breadcrumb
=> the thumbnail is not up-to-date with the new background color.
You need to reload the page to see it.

Technical note: for regular documents, `unique` is the attachment checksum.
It doesn't work for spreadsheets since the attachment actually doesn't change
when a spreadsheet is updated. A revision is dispatched.
The last revision uuid can be used as `unique` discriminant instead

closes odoo/enterprise#61293

Task: 3887916
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>