[FIX] account_reports: journal report audit tax grid
Before this commit, when having opposite tax grid the domain of the
journal_report_tax_tag_template_open_aml function was wrong.

In case of opposite tax grid we should have put a list of tax_grid in the tag_id
key which wasn't the case. It was only taking the last one since it was
overridden in the next iteration of the loop.

task-3869482

closes odoo/enterprise#62407

X-original-commit: 9a3ce806c1cc510545fc4e7d379209cc404a66d5
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>