[FIX] account_accountant: fix transaction details display bank rec widget
In saas-16.3 we started to use a dedicted field transaction_details on statements [1]
instead of the narration fields from `account.move`.

1. Fix the view so that it displays the right field.
2. Move transaction details to the same place regardless of the reconciled
status of the transaction, into the dedicated tab.
Previously, this transaction details was displayed in the manual operations tab
when the statement was not reconciled and in debug mode.
When reconciled it was available in a dedicted tab "Transaction Details" (no need to be in debug).

[1] https://github.com/odoo/odoofin/blob/17.0/odoofin_proxy/models/odoofin_proxy.py#L315

closes odoo/enterprise#61734

Task-id: 3888162
X-original-commit: 3bb2fa7944d296692c6858fa9661bc54bcd94659
Related: odoo/upgrade#5973
Signed-off-by: Laurent Smet (las) <las@odoo.com>