[IMP] l10n_pe_edi: message in the chatter
Before this commit, when you submit an invoice through the EDI even in testing
mode, the message in the chatter is saying that the document has been sent to
the government. Which is disturbing for the users, some of them think they
submitted correctly the invoice but in fact it was in testing mode.

From now on, when the setting is set, the message is changed to say that you are
 in testing mode.

task-3300375

closes odoo/enterprise#61318

Signed-off-by: William André (wan) <wan@odoo.com>