[FIX] web_studio: fix failing test
The test was failing on this step:

```js
{
    // checks that the field doesn't exist anymore
    extra_trigger: 'label.o_form_label:not(:contains("COUCOU"))',
    trigger: ".o_web_studio_leave > a.btn",
}
```

Before https://github.com/odoo/enterprise/pull/60009 ,
restoring the view didn't work.

The extra_trigger was meant to check that no label containing "COUCOU"
was present on the view, but what actually happened is that the
extra_trigger matched other labels.

With the previous fix, when the view is restored, there is no more
label, thus the step failed.

closes odoo/enterprise#61706

X-original-commit: 752ff5449fb47c69e0050d7a70fd6d509a44940c
Related: odoo/odoo#163887
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>