[IMP] l10n_ke_hr_payroll: adding secondary employee
In Kenya, secondary employees are taxed at a flat rate of 30%.
They are not eligible to any reliefs and tax exemptions (Tax,
Insurance & Mortgage Relief), and pension does not reduce
taxable income.

What's a second employee ? it's a contract that is not for
your main company. We can imagine you work in a company A
(managed by Odoo or not) where your usual PAYE rate will apply
and you have a second (or third, or else) contract in others
company, for "not-primary" job.

task-3999125

closes odoo/enterprise#67128

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>