[FIX] industry_fsm: use correct planned_date_begin at task creation
Currently, when creating a task from the calendar view in the field service app,
the start date that was set by the user is not registered.

Steps to reproduce:
-------------------
* Go to the **Field Service** App
* Switch to the calendar view
* Select "Tomorrow" (x+1)
* For the new task, set the `Planned Date` as follows:
  * Start date : The day after tomorrow at 3PM (x+2 3PM)
  * End date: Three days after the start at 4PM (x+5 4PM)
* Fill anything for the other required fields > **Save & Close**
* Switch to the kanban view and select the task created
> Observation: the start date of `Planned Date` is set at Tomorrow 7AM instead of the day after at 3PM

Why the fix:
------------
This behavior was introduced after this commit:
https://github.com/odoo/enterprise/commit/39b9362d0469d71b59d47e9d66d5e73b9a46520b.

This commit added an inverse method to the field `planned_date_start` to enable
drag and droping tasks in the calendar view. And `planned_date_start` is a field
that exists purely for calendar displaying of tasks.

https://github.com/odoo/enterprise/blob/c8c4a52b508621b45239332273832de26a085931/project_enterprise/models/project_task.py#L299-L305

When creating a task in the calendar view, the context contains `default_planned_date_start`
and its value depends on the first day you selected in the calendar. At task creation,
`default_planned_date_start` triggers a write in `planned_date_start` which triggers its
inverse method.

It does not make sense to overide the user choise for `planned_date_begin` with
`default_planned_date_start` when creating a task. We revert the js change made
in this commit: https://github.com/odoo/enterprise/commit/39b9362d0469d71b59d47e9d66d5e73b9a46520b

This removes the `default_planned_date_start` from the context which won't trigger the
inverse method. We don't have to worry about the line
`default_planned_date_begin: default_planned_date_start,` as if we have selected a start
date in the form, it will be present inside `vals_list` in the create function and will
use that value instead of the one in the context.

opw-3850455

closes odoo/enterprise#62301

X-original-commit: 7606352004f37e39e4713a4ddc371f1e2252763d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>