[IMP] sale_planning: edit planning hours fields filter
Currently, the fields 'planning_hours_planned' and
'planning_hours_to_plan' are calculated by filtering out SOLs that don't
have planning enabled or that are in the draft or sent state, and by
converting the quantity of these SOL to hours according to the unit. The
problem with this is that if we manually choose an SOL that doesn't have
planning enabled in the planning slot, it does not count in the progress
bar.

This PR changes the filters to include SOLs that have planning disabled.

task-3857649

closes odoo/enterprise#61326

Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>