[FIX] delivery_sendcloud: fix `None` functionality filter
Steps to reproduce:
Configure Sendcloud with a shipping product that provides
customizable functionalities, and select `None` as the
functionality filter. All of the shipping methods will
be filtered out and we get an error message:
`There's no shipping method matching all your selected
filters for this picking/order.`

The issue is that `'None'` is passed as a string in the list
of functionalities but it is checked as `None` type.

To fix the issue we check if the string exists in the filter.

opw-3878116

closes odoo/enterprise#61360

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>