[FIX] documents_spreadsheet: correctly simulate a click
Before this commit, the test was not correctly simulating a click on the
link chart button. This was due to the fact that the only event that was
triggered was the pointerdown event, which is not enough to trigger the
click event.

For an unknown reason, only triggering the pointerdown event has a side
effect during the test to trigger an event handled in the Leatlef library,
would lead to an error in the library, and thus in the test suite.

This commit fixes the issue by using the click helper function from the
web client test helpers.

closes odoo/enterprise#61430

Runbot: 59203
X-original-commit: ad3906c407dcdb25d34e64c35e5e29e1a7f6b45f
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Co-authored-by: Rémi Rahir <rar@odoo.com>
Co-authored-by: Pierre Rousseau <pro@odoo.com>