[IMP] project_enterprise: add embedded actions
This commit is part of a task implementing embedded actions for the module
project. This task also imaed to remove the project update status from
the control panel override made in project. The changes of this commit
aim to remove the control panel override from the different views.

taskid:3750721

Part-of: odoo/enterprise#61388