[FIX] account_reports: fix groupby on non-relational fields of account.move.line
Trying to group on a field such as the boolean tax_tag_invert, for example, used to crash because of the value conversion that was done in the process of parsing the generic id assigned to the groupby line.

We fix that by ensuring the value only gets converted to an int when a model is assigned to the line, meaning the value has to be a model id. In all other cases, the value will be put as-is in the generic id, and its string representation will hence be used in the UI.

closes odoo/enterprise#62002

X-original-commit: c840d19579427e01ef68de82d2768c1eea3319f7
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>