[REF] l10n_mx_edi_extended: hooks to standard csv file
Previously, in l10n_mx_edi_extended, the following objects:

- `res.city`
- `l10n_mx_edi.res.locality`
- `l10n_mx_edi.tariff.fraction`

They was previously made in the post_init_hook instead of a standard csv
file for optimization reason. This is a bit incovenient, because those
objects can't be referred inside the same module to be a demo data
(because they are not created yet).

This commit convert them to the standard csv format. The performance is
slightly worse (from 5.73s to 7.24s) but in return we get:

- easier to look for when searching (now we have the full id written)
- no more intellisense errors
- less confusing code
- more general/standard approach

Task-id: 3851842
Part-of: odoo/enterprise#60006