[FIX] voip: add back customer info on calls
Functionality was removed during the refactor of the voip module between
16 and 17 where you could enter any number through the phone keypad on
voip or receive a call and have access to the information on the
customer with the associated phone number.

This was a functionality that the customer on the ticket was using
prior to upgrade.

I have added back this functionality through a non blocking call to a
new function in voip.call that will search up the contact based on the
phone number entered in the softphone. When the data comes back from the
back end the UI will be updated with the partner information if it found
it.

This also has a side effect of fixing the customer wizard button as
well.

opw-3770625

closes odoo/enterprise#60383

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>