[FIX] documents: fix unscrollable inspector
Purpose
=======
Prevent the horizontal scrolling of the inspector but allow
the vertical scrolling for small screen sizes.

Specifications
==============
Reverting the addition of the "overflow-hidden" class on the
documents inspector as it is preventing its horizontal but also
its vertical scrolling. On lower screen sizes, a vertical scrolling
is needed or else the inspector becomes pratically unusable.

related PR: odoo/enterprise#59652

Task-3884149

closes odoo/enterprise#61436

X-original-commit: ebf86aef7e11312325c00172c7d6b8b69faf730d
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Amélie Dieudonné (amdi) <amdi@odoo.com>