[FIX] appointment: fix gantt domain
In [1] we added a domain on appointment actions to prevent a traceback
when users visit the gantt appointment views when resources from
different companies appear at the same time.

However appointment_resource_id was deprecated in favor of
appointment_resource_ids in stable since then.

`appointment_resource_ids` should now be used in the domain.

[1]: ab7addece528d887f062e70732f96fde2ed76dca

task-3893256

Part-of: odoo/enterprise#61439