[FIX] account_intrastat/l10n_be_intrastat: Grouped lines for BE only
In this commit [1], we changed the way we display intrastat report lines
by grouping them. We did it for all the intrastat report, it was wrong.
Actually, we should only do it for the belgian report.
The aim of this commit is adding a key in options to know if lines should
be grouped or not.
By default the option is set to False except for the Belgian report.
Since version 16.4, users can change the value by using the filter.

task-3892823

[1]: https://github.com/odoo/enterprise/commit/217594fe8ba00329b3eb5c68e3eca3b556127cde

closes odoo/enterprise#61538

X-original-commit: 3c5bc50d84ef6ad7ec23000a28936687e667a1fa
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>