[IMP] l10n_din5008_industry_fsm: xml instead of python
A lot of information of the din5008 layout were computed with some compute
function which made the customization a bit hard to do.
This commit will modify the l10n_din5008 module to pass all the logic in xml.

Also updating the translation of the module.

task-3888498

closes odoo/enterprise#61458

Related: odoo/odoo#160689
Signed-off-by: William André (wan) <wan@odoo.com>