[IMP] website_{appointment,enterprise}: add appointments dynamic snippet
In order to easily display existing appointments on the website, add an 'Appointments'
dynamic snippet. It will display punctual and recurring appointments, published and
linked to the current website (or all).

Also, a few options are added:
- One can filter on Users / Resources
- If one of those is used, the user can select existing users / resources to
filter the displayed appointments
- Also, one can specify a list of comma separated names that will be used to
search appointments based on 'name'. Only non empty trimmed values are used..

Some placeholder data is added. The snippet template is the same one as appointment
cards on the front-end. The main difference is the display of the sample ribbon.

Replace o_object_fit_cover by object-fit-cover, as it will be necessary to have an
'cover !important' obect-fit value, to override s_dynamic default scale-down value.
For the appointment cards, we want the object-fit set to cover, always.

The previous 'Appointment' snippet, which adds a button that can redirect to
appointments (with some options) is renamed (in strings, not in the files, to
avoid version issues) to 'Appointment Button' to be more explicit.

Task-3889948

Part-of: odoo/enterprise#61471
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>