[FIX] knowledge: various record usage fixes
Topbar:
`record.update` marks the record as dirty and in need of a `save`, but the
change was already saved by `action_set_lock/unlock`. Switching lock mode should
not mark the record as dirty as it will wrongly change the display of the
`FormStatusIndicator`.

Sidebar:
The callback of `useRecordObserver` will be called only if a `record` property
is set to a different value AND was previously read (aka Proxy subscription).
When the sidebar was loading but did not already previously load the current
article (aka first load), no property of the record was read, so any change to
the first article was not impacted in the sidebar until the next load. (i.e.
renaming, changing the icon).
A console error log is removed because it is better to display a proper error
message instead of hiding it in the console, and it would fail tests either way.

Knowledge Cover:
`onWillUPdateProps` should not be used to detect `resId` changes, since in some
cases its callback won't be called. (i.e. `Record:save({nextId})`)

Comments Handler:
An asynchronous callback for `useRecordObserver` will not delay a pending
Component patch. Therefore a `onPatched` callback can not depend on a value set
asynchronously in the `useRecordObserver` callback, since there is no guarantee
that it will be ready in time. To remediate the issue before the comments
refactoring, a `KeepLast` is introduced to be able to delay processing depending
on the result of the `useRecordObserver` callback to ensure that everything is
done relative to a valid state of the application.

task-3875521

Part-of: odoo/enterprise#60960
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>