[FIX] gantt_model: correct filter groupby
Issue:
------
The commit[^1] filters groupby in the gantt view according
to whether it is a propertie or not.
However, fields of type `date` and `datetime` use `:` to refer
to the granularity (example `check_in:year`).
To avoid const { type } = metaData.fields[fieldName:granularity]; to
trigger an error, it is therefore necessary to adapt the diff of
commit[^1].

opw-3893042

[^1]: https://github.com/odoo/enterprise/commit/6ee5fe8220d692173187402a7361f77883dca921

closes odoo/enterprise#61522

Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>