[IMP] documents: add color picker for tags category
Earlier, a group color was computed for tags based on the order of facets.

This commit adds a color picker widget in the `tag category edition modal`,
as well as in the `tags configuration form`.

If a group color is set by the user for a tag category, it is now directly
applied to all the tags belonging to that category, instead of performing any
sort of computations. Else, a randomly generated default tag color is applied.

Task-3874977

Part-of: odoo/enterprise#61598