[REF] point_of_sale: unify test setup
The test assets that we use in pos are duplicated across different
classes. In the corresponding commit we take the first steps in unifying them, by
having the class in `pos_restaurant/tests/test_frontent.py` inherit from
`point_of_sale`.

In this commit we make the necessary changes to the enterprise modules.

Task 3895255

closes odoo/enterprise#61529

Related: odoo/odoo#155540
Signed-off-by: David Monnom (moda) <moda@odoo.com>