[FIX] l10n_eu_oss_reports: fix the display of column groups within the OSS report
Before, using a column group on the OSS report broke the layout, as each country line displayed more cells than it should, and they hence mismatched the column headers.

This was due to bad management of the column groups: when there was more than one column group, the tax_sums variable contained more than one value, and we were then looping on each column for each tax sum to generate the columns of the country line, which was wrong, since options['columns'] already contained the columns of each column group.

closes odoo/enterprise#61530

Signed-off-by: John Laterre (jol) <jol@odoo.com>