[FIX] stock_barcode_mrp: don't force reservation
Issue:
===============
When a manufacturing order (MO) is created, the default components are
automatically reserved and are made visible in the Barcode app.
However, even if these components are later unreserved, they will
continue to be displayed in the barcode module.
The issue is: if the MO was not confirmed, it is automatically reserved
when opening it in the Barcode app. And so, since components moves are
reserved again, they are shown in the Barcode app.

Resolution:
================
To address this issue, we've taken the step to avoid forcefully re-reserving
the components when a component is unreserved. This resolves the persistent
display of components in the barcode.

Steps to Reproduce:
======================
1. Create a manufacturing order (MO) and confirm it.
2. Navigate to the barcode module within the manufacturing order.
3. By default, the components are reserved, so they are visible in the
   barcode module.
4. Go to the form view of the MO in MRP and unreserve the components.
5. Check the barcode module again, and notice that the unreserved components
   are still visible.

Expected Result:
==================
After implementing the solution, when we opens the MO in the Barcode
app, we should not automatically reserve its components.

task-3869731

X-original-commit: a4db83136f5632aca57dc7d73196cb5425912c5c
Part-of: odoo/enterprise#67142
Signed-off-by: Steve Van Essche <svs@odoo.com>