[FIX] helpdesk: prevent traceback when setting the team_id of a ticket to False
When the team_id of a ticket is set to False in its form view, a
traceback occurs. This is an invalid value anyway, but the user should
only be notified of that fact when he tries to save his change, not
with traceback.

Step to reproduce:
- Open helpdesk
- open the 'all tickets' menu
- open any ticket form
- set the team_id to false
- save changes or click anywhere else to leave the edit field
=> a traceback occurs

Source of the issue:
When a new team is set on a ticket, a new stage is set on it, as well as
new sla_status. The issue is that inside the _compute_sla_deadline
method, the calendar of the team is used. But since there are no
teams, there are also no calendar. And the ensure_one() fails later on
in the stack because of that.

Solution:
Prevent the computation of those values for ticket without team_id.
Since this is an invalid value, the changes would be rolled back anyway,
or overwrite once the user put a valid value for the team.

version 16.0 - master
task-3895303

closes odoo/enterprise#63039

X-original-commit: bbd576d22dc57bde63c810a828fc1f1bde98a55c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>