[IMP] account_accountant: add reco models to generate invoice/bill
As of now, it is not possible to create a commercial move directly from the
bank reconciliation widget. This commit aims to facilitate the workflow to
create invoices or bills and reconcile them with existing bank statement lines.

---
Current behaviour:

Before this commit, if the user has a bank statement line for a unregistered
commercial event, they have to:
1. Make a note of the amount, date and potentially a label
2. Navigate to Customers/Vendors > Invoices/Credit Notes/Bills/Refunds > New
3. Go back to the reconciliation widget (no breadcrumbs available)

---
Expected behaviour:

After this commit, the user will be able to create a move through a button
in the bank reconciliation widget. The move is pre-filled with relevant
information, such as amount, date and partner (if available). The user can
return to the bank reconciliation widget through the breadcrumbs, and proceed
to validate the created and added move.

---
This commit:
- Modifies behaviour of js function actionSelectRecoModel to support the creation
of moves.
- Modifies function _action_select_reconcile_model to continue supporting
journal entries as previously, but also the creation of moves through new functions
_create_invoice_from_write_off_values and _get_invoice_price_unit_from_price_total.
- Modifies _js_action_restore_st_line_data to add the created move to the
reconciliation widget.
- Adds tests for the creation of moves through the reconciliation button.

task-3768742

closes odoo/enterprise#59555

Related: odoo/odoo#159502
Signed-off-by: Laurent Smet (las) <las@odoo.com>