[FIX] account_invoice_extract: fix "Reload AI data" with a matched PO
When a purchase order was matched and imported by the OCR, it wasn't
working properly, the data imported was overwritten with the detected
fields of the OCR.

This happened because, in the implementation before this commit, a flag
`force_write` was used to indicate that the data should be overwritten,
but this implied that the data imported from the PO was also
overwritten.

In this new implementation, the content of the invoice is first reset
and then the logic to fill in the invoice with the OCR data and/or
purchase order is applied.

closes odoo/enterprise#63542

Task: #3834665
X-original-commit: 8fb17756794f18cbd632b4b231fd49d9674970db
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>