[FIX] l10n_be_reports_post_wizard: show vat export comment in XML
When generating VAT XML by posting a tax closing entry, the
`<ns2:Comment>` tag in the ML is empty despite specifying a comment
during the closing process.

Steps to reproduce
------------------
1. Install `l10n_be_reports_post_wizard`.
2. Switch to a Belgian company.
3. Open the tax report.
4. Create a tax closing.
5. Post the tax closing, ensuring a comment is defined in the appearing
   wizard.
6. Review the generated XML file.

Expected behavior: The `<ns2:Comment>` tag in the XML should display the
comment entered during the tax closing.

Actual behavior: The `<ns2:Comment>` tag is found to be empty.

opw-3859928

closes odoo/enterprise#61506

X-original-commit: b4af309d2b3c3514520c84a9ec79612edefe3d3c
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>