[IMP] l10n_syscohada_reports: add reports for associations
Adding another report for SYSCEBNL, the one for Associations.

All the COAs of countries from SYSCOHADA are children from the syscohada template.
So, we create a new availability for children of a COA, as we don't want Companies to have access to the reports of an Association (and vice-versa).

task-3456480

closes odoo/enterprise#67434

X-original-commit: 733559897cabb2562186feb9bc2df2bc16ca99d0
Related: odoo/odoo#174646
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>