[IMP] l10n_syscohada_reports: add reports for associations
Adding another report for SYSCEBNL, the one for Associations.

All the COAs of countries from SYSCOHADA are children from the syscohada template.
So, we create a new availability for children of a COA, as we don't want Companies to have access to the reports of an Association (and vice-versa).

task-3456480

closes odoo/enterprise#62883

Related: odoo/odoo#166211
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>