[FIX] hr_payroll: removed the duplicate menu action
Before Commit:

Prior to this update, in the 'list_controller.js' file in version saas~16.4 did
not included the Duplicate menu action within it. However, in the same
version, the 'hr.payroll' model defined this action in the
'hr_payroll_structure_views.xml' file, utilized within the 'Salary Structures'
in the Payroll app. Now, in version 17.0, this Duplicate menu action was
added to the 'list_controller.js'.

After Commit:

Following the update, despite the addition of the Duplicate menu action in the
'list_controller.js' file in the version 17.0, the very menu action still
remained present in the particular payroll module. Consequently, we have
removed it, to ensure the same action do not exist twice.

Task ID: 3854557

closes odoo/enterprise#61666

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>