[FIX] project_timesheet_forecast: update effective hours
Steps to reproduce:
-------------------
- create a project;
- create a new planning slot for the project with a resource (user);
- go to the project and create a task;
- record timesheet in the planning slot time range in the task;
- go to the planning slot view (via list view)

Issue:
------
The times recorded are not updated (in the smart button).
Despite the fact that `_compute_effective_hours` depends
on `timesheet_ids`.

Cause:
------
The `effective_hours` field will be recomputed if one of
its dependencies changes (in this case, `timesheet_ids`).
For `timesheet_ids` to be marked for compute,
one of its dependencies must change.

`project_id.analytic_account_id.line_ids`
is different from `project_id.timesheet_ids`.

As a result, the method which computes `effective_hours`
is not triggered.

Solution:
---------
Add `project_id.timesheet_ids` in dependencies.

opw-3891624

closes odoo/enterprise#61668

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>