[FIX] account_reports: consider hide lines at 0 when printing
0 lines are still included in exported report when Hide lines at 0 is toggled

Steps to reproduce:
- Open a report, e.g. Balance Sheet (Accounting > Reporting > Balance Sheet)
- In options, toggle 'Hide lines at 0'
- Click in PDF to export a PDF version of the report
- 0 lines are included in the PDF file

When "Hide lines at 0" is toggled, lines at 0 are still included in the PDF/XLSX
report, despite being hidden in the report view. This leads to a difference
between what the user sees and what is printed.

Another reason to hide lines at 0 from the printed report is that lines with Hide
if Zero checked does impact the PDF. This means the two options (Hide lines at 0
and Hide if Zero) have similar impacts in the user view but different behavior in
printed reports.

This commit implements the function _filter_out_0_lines to remove lines at 0 from
printed report if "Hide lines at 0" is toggled, and adds a test.

task-3888290

closes odoo/enterprise#63230

X-original-commit: 612bedefe673468d15edae87b089d74f3fb000cf
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Maira Salazar <srma@odoo.com>