[FIX] website_helpdesk_knowledge: add discardForm during macro tour
The `send as message` macro for the template/clipboard block is only supposed to
open and edit the message composer on the target form view, then discard it. The
form view does not become dirty because of that.

However it has been observed that on some Odoo versions, the form view sometimes
becomes dirty (indeterministic factor), which is unrelated to the successful
macro operation.

Therefore it has been decided to always attempt to discard the form so that the
test does not fail no matter if the view is dirty or not during the test.

runbot error 62459

closes odoo/enterprise#61759

X-original-commit: 58dae7d92d8d95f4e36bbab5db2dd9f3abd91ee7
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>