[IMP] stock_barcode: show quantity to count
Add a new setting, "Show Quantity to Count", who is set by default.
When set, quantity to count will be displayed on the Barcode Inventory
Adjustment, just like before this commit.
When the setting is off, the quantity to count will not be displayed.

The purpose is to make it not too easy for inventory workers to "count"
the products. Because when the expected quantity to count is displayed,
the user can simply hit the button to mark the quantity as counted.
With the expected quantity to count hidden, the inventory workers have
no choice but really count the the products.

task-3479636

Part-of: odoo/enterprise#49397
Co-authored-by: svs-odoo <svs@odoo.com>