[IMP] stock_barcode: count entire locations
Adds a new setting, "Count Entire Locations".
When this setting is on:

1. Scanning a location in the Barcode app Inventory Adjustment will
   fetch all scanned location's quant.

2. When requesting a count on some quants, all other quants sharing the
   same location(s) will be set a to count too.

The purpose of this setting is to avoid to miss some products while
counting products after starting a fresh inventory adjustment in the
Barcode app while not previously set manually the quants to count.

task-3479636

closes odoo/enterprise#49397

Related: odoo/odoo#154848
Signed-off-by: Steve Van Essche <svs@odoo.com>
Co-authored-by: svs-odoo <svs@odoo.com>