[FIX] appointment: set condition based on type
Added condition on a check in microsoft calendar that ensured the organizer
was an attendee on the event on create. This was added in a change that allowed the organizer to be
changed on the Odoo side. However, this caused issues with the
appointments app when creating an appointment that only used resources.
The "organizer" who in this case is the creator of the appointment type
would not be an attendee and therefore would cause an error and make the
appointment type unbookable.

Adding this condition allows the appointment to bypass the check if the
appointment type uses resources instead of users.

This change adds the necessary override in enterprise. See change in
community here: https://github.com/odoo/odoo/pull/162927/commits/285eb5a6efc8fd6bbbc301458a6a0e16cad7efc1

opw-3841495

closes odoo/enterprise#61731

Related: odoo/odoo#162927
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>