[FIX] web_studio: disambiguate view diffing during normalization
Steps to reproduce
==================

- Open a product form
- Open Studio
- Remove the "Internal Notes" block from "General Information"
- Add a new field

=> The normalized view contains many operations, it should only contain
one move and one insertion

Cause of the issue
==================

The normalization compares the original view with the view with all
operations applied. To do this, it stringifies every node and then does
a textual diff between the original and modified versions.

The issue in this case is that there are multiple ways to represent a
diff.

In this example, we remove the first group

```xml
<form>
    <group/>
    <group/>
</form>
```

```xml
<xpath expr="/form/group" position="replace"/>
```

The before stringified view is represented as

```
/form/group
/form/group[@closed]
/form/group
/form/group[@closed]
```

The after stringified view is represented as

```
/form/group
/form/group[@closed]
```

There are two ways to represent this diff

```diff
- /form/group
- /form/group[@closed]
  /form/group
  /form/group[@closed]
```

```diff
  /form/group
- /form/group[@closed]
- /form/group
  /form/group[@closed]
```

In this case, we only want the first one, but the second one is
generated.

Solution
========

We can assign an id to every node before performing the diff, this way,
the nodes are uniquely represented.

The before stringified view will be represented as

```
/form#1/group#2
/form#1/group#2[@closed]
/form#1/group#3
/form#1/group#3[@closed]
```

The after stringified view is represented as

```
/form#1/group#3
/form#1/group#3[@closed]
```

The diff can now only be represented in one way

```diff
- /form#1/group#2
- /form#1/group#2[@closed]
  /form#1/group#3
  /form#1/group#3[@closed]
```

---

Backport of 49d6e0fdceae1fb57378e87d93de5751f9210d16

opw-3819667

Part-of: odoo/enterprise#67980
Related: odoo/odoo#175867
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>