[FIX] l10n_ec_reports_ats: filter out cancelled draft moves in ATS report
Steps to reproduce:
- Install l10n_ec_reports_ats
- Switch to an Ecuadorian company (e.g. EC Company)
- Create a draft invoice
- Cancel the invoice
- Go to "Accounting / Reporting / Statement Reports / Tax Report"
- Filter the report on the same month than the cancelled invoice
- Click on upper-left ATS button

Issue:
A traceback is raised:
"TypeError: 'bool' object is not subscriptable"
while trying to render the cancelled moves:
<secuencialInicio t-out="void_move.l10n_latam_document_number[-9:]"/>

Cause:
"l10n_latam_document_number" field is a computed field that depends on
the name of the account move.
When the move is in draft, it doesn't have a name (i.e. "/") and
"l10n_latam_document_number" has False as value.

Solution:
Cancelled draft invoices should not be taken into account in ATS report.

opw-3857645

closes odoo/enterprise#61901

X-original-commit: 52a570b7cf10d2bdd64018fee04e3c204671cc6c
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>