[FIX] spreadsheet_edition: compute current revision uuid
Invoking `_get_spreadsheet_snapshot` with `bin_size=True` in the context
crashes if the spreadsheet has been snaphotted. That's because
we are trying to json parse the human readable size version of the field
'10.5 ko', which is obviously not a valid json.

Since a055dfeb84abc, the computed field `current_revision_uuid` is in
the view (which always has `bin_size=True`). And the compute method
calls `_get_spreadsheet_snapshot`

closes odoo/enterprise#62540

Task: 3899662
X-original-commit: a6d69bbf1cd5afedcf7dae3d460d05211ab659aa
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>