[FIX] sale_subscription: display stages data in kanban view
Issue:
======
In kanban view , some data are missing.

Steps to reproduce the issue:
=============================
- Make sure to have some subscriptions in progress stage.
- Go to kanban view, there is no subscription in the progress stage.

Origin of the issue:
====================
Calling `_group_expand_states` will add 2 stages `3_progress` and
`4_paused` but when calling `web_read_group` there is already group
`3_progress` so in `read_group_fill_result`  we will have duplicate
group, The first we loop over the group we have the data correctly but
we remove it from the dict and then for the second time we don't find
any data and it will be overriden and become empty.

Solution:
=========
Return only distinct groups.

opw-3891031

closes odoo/enterprise#61817

X-original-commit: 2d56866057112af310ba2c22dd939e12310d3dc2
Signed-off-by: Mahdi Cheikh Rouhou (macr) <macr@odoo.com>