[FIX] account_winbooks_import: don't reconcile 0 balance items
The reconciliation of journal items imported from Winbooks is done
when all those items are posted. However if one of the imported items
to reconcile have zero as balance, it will trigger an error
"You are trying to reconcile some entries that are already reconciled."
because Odoo consider zero balance lines as reconciled and block the user of
posting the entries.

This fix removes the reconciliation data when the Winbooks file is imported
if the item has a zero balance since this item does not need to be reconciled
in the future.

opw-3830355

closes odoo/enterprise#61743

Signed-off-by: William André (wan) <wan@odoo.com>