[FIX] web_studio: approvals: granting should only notify applicable higher rules
Have rules such as:
1. no domain, notification order: 1
2. no domain, notification order: 2
3. a domain that doesn't match the record, notification order: 2

Manually validate the rule 1, that is, do not hit the button of the action, rather, enter the popup
of the approvals, and validate the rule.

The spec of the flow is that, when all rules from a level (or notification order) have been validated, rules
that have a higher order will trigger an activity and a notification to their responsible. After all, those rules are
"next in line".

Before this commit, both rule 2 and 3 sent a notification and created an activity for the current record.
This was wrong as rule 3 should not be considered, because its domain doesn't match the record.

After this commit, only rule 2 sends a notification and creates an activity.

opw-3815732

closes odoo/enterprise#61746

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>