[IMP] sale_renting: show rental period in calendar view
Before the commit:
The calendar view of the rental was based on the next activity date (which is
expected for sales and subscriptions). Customers expect to see the orders based
on the rental period. There was no customized calendar view dedicated to rental
orders.

After the commit:
Now the rental periods are shown in the calendar view which is based on the
rental start date and rental return date.  The new view is extended from the
calendar view of the sale order and overridden necessary attributes to display
the rental period

task-3893588

closes odoo/enterprise#61760

Related: odoo/odoo#165056
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>