[IMP] hr_appraisal: enable indirect managers to request employee appraisals.
The system now allows for creating appraisals for
individuals even if you're not their direct
manager but are in their managerial hierarchy.
For instance, if A is the manager of B, and B is
the manager of C, A can request an appraisal for C.
Previously, this resulted in an error because only
direct managers could create an appraisal.

task-3890348

closes odoo/enterprise#61764

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>